fix(mcp): retain final disconnect across accept iterations
